requirements.txt
# python=3.8.16
matplotlib=3.7.1
numpy=1.23.5
pillow=9.4.0
pip=23.0.1
setuptools=65.6.3
numpy==1.23.5
matplotlib==3.7.1